+++ discord/guild.py
@@ -35,7 +35,7 @@ from .emoji import Emoji
 from .errors import InvalidData
 from .permissions import PermissionOverwrite
 from .colour import Colour
-from .errors import InvalidArgument, ClientException
+from .errors import InvalidArgument, ClientException, InsufficientPermissions
 from .channel import *
 from .enums import VoiceRegion, Status, ChannelType, try_enum, VerificationLevel, ContentFilter, NotificationLevel
 from .mixins import Hashable
@@ -803,6 +803,8 @@ class Guild(Hashable):
             pass
 
         parent_id = category.id if category else None
+        if not self.me.guild_permissions.manage_channels:
+            raise InsufficientPermissions('manage_channels')
         return self._state.http.create_channel(self.id, channel_type.value, name=name, parent_id=parent_id,
                                                permission_overwrites=perms, **options)
 
@@ -945,6 +947,8 @@ class Guild(Hashable):
         HTTPException
             Leaving the guild failed.
         """
+        if self.owner.id == self.me.id:
+            raise InsufficientPermissions('you_are_owner')
         await self._state.http.leave_guild(self.id)
 
     async def delete(self):
@@ -960,7 +964,8 @@ class Guild(Hashable):
         Forbidden
             You do not have permissions to delete the guild.
         """
-
+        if not self.owner.id == self.me.id:
+            raise InsufficientPermissions('you_are_not_owner')
         await self._state.http.delete_guild(self.id)
 
     async def edit(self, *, reason=None, **fields):
@@ -1051,6 +1056,8 @@ class Guild(Hashable):
         except KeyError:
             pass
         else:
+            if not self.me.guild_permissions.manage_guild:
+                raise InsufficientPermissions('manage_guild')
             await http.change_vanity_code(self.id, vanity_code, reason=reason)
 
         try:
@@ -1120,6 +1127,8 @@ class Guild(Hashable):
             raise InvalidArgument('system_channel_flags field must be of type SystemChannelFlags')
 
         fields['system_channel_flags'] = system_channel_flags.value
+        if not self.me.guild_permissions.manage_guild:
+            raise InsufficientPermissions('manage_guild')
         await http.edit_guild(self.id, reason=reason, **fields)
 
     async def fetch_channels(self):
@@ -1262,6 +1271,8 @@ class Guild(Hashable):
         BanEntry
             The BanEntry object for the specified user.
         """
+        if not self.me.guild_permissions.ban_members:
+            raise InsufficientPermissions('ban_members')
         data = await self._state.http.get_ban(user.id, self.id)
         return BanEntry(
             user=User(state=self._state, data=data['user']),
@@ -1293,7 +1304,8 @@ class Guild(Hashable):
         List[BanEntry]
             A list of BanEntry objects.
         """
-
+        if not self.me.guild_permissions.ban_members:
+            raise InsufficientPermissions('ban_members')
         data = await self._state.http.get_bans(self.id)
         return [BanEntry(user=User(state=self._state, data=e['user']),
                          reason=e['reason'])
@@ -1343,7 +1355,8 @@ class Guild(Hashable):
 
         if not isinstance(days, int):
             raise InvalidArgument('Expected int for ``days``, received {0.__class__.__name__} instead.'.format(days))
-
+        if not self.me.guild_permissions.kick_members:
+            raise InsufficientPermissions('kick_members')
         data = await self._state.http.prune_members(self.id, days, compute_prune_count=compute_prune_count, reason=reason)
         return data['pruned']
 
@@ -1364,7 +1377,8 @@ class Guild(Hashable):
         List[:class:`Webhook`]
             The webhooks for this guild.
         """
-
+        if not self.me.guild_permissions.manage_webhooks:
+            raise InsufficientPermissions('manage_webhooks')
         data = await self._state.http.guild_webhooks(self.id)
         return [Webhook.from_state(d, state=self._state) for d in data]
 
@@ -1397,7 +1411,8 @@ class Guild(Hashable):
 
         if not isinstance(days, int):
             raise InvalidArgument('Expected int for ``days``, received {0.__class__.__name__} instead.'.format(days))
-
+        if not self.me.guild_permissions.kick_members:
+            raise InsufficientPermissions('kick_members')
         data = await self._state.http.estimate_pruned_members(self.id, days)
         return data['pruned']
 
@@ -1421,7 +1436,8 @@ class Guild(Hashable):
         List[:class:`Invite`]
             The list of invites that are currently active.
         """
-
+        if not self.me.guild_permissions.manage_guild:
+            raise InsufficientPermissions('manage_guild')
         data = await self._state.http.invites_from(self.id)
         result = []
         for invite in data:
@@ -1523,6 +1539,8 @@ class Guild(Hashable):
         img = utils._bytes_to_base64_data(image)
         if roles:
             roles = [role.id for role in roles]
+        if not self.me.guild_permissions.manage_emojis:
+            raise InsufficientPermissions('manage_emojis')
         data = await self._state.http.create_custom_emoji(self.id, name, img, roles=roles, reason=reason)
         return self._state.store_emoji(self, data)
 
@@ -1611,12 +1629,19 @@ class Guild(Hashable):
         for key in fields:
             if key not in valid_keys:
                 raise InvalidArgument('%r is not a valid field.' % key)
-
+        if not self.me.guild_permissions.manage_roles:
+            raise InsufficientPermissions('manage_roles')
         data = await self._state.http.create_role(self.id, reason=reason, **fields)
         role = Role(guild=self, data=data, state=self._state)
 
         # TODO: add to cache
         return role
+
+    def can_kick(self, user):
+        if not self.me.guild_permissions.kick_members:
+            raise InsufficientPermissions('kick_members')
+        if self.me.top_role < self.get_member(user.id):
+            raise InsufficientPermissions('kick_members_hierarchy')
 
     async def kick(self, user, *, reason=None):
         """|coro|
@@ -1642,7 +1667,14 @@ class Guild(Hashable):
         HTTPException
             Kicking failed.
         """
+        self.can_kick(user)
         await self._state.http.kick(user.id, self.id, reason=reason)
+
+    def can_ban(self, user):
+        if not self.me.guild_permissions.ban_members:
+            raise InsufficientPermissions('ban_members')
+        if self.me.top_role < self.get_member(user.id):
+            raise InsufficientPermissions('ban_members_hierarchy')
 
     async def ban(self, user, *, reason=None, delete_message_days=1):
         """|coro|
@@ -1671,6 +1703,7 @@ class Guild(Hashable):
         HTTPException
             Banning failed.
         """
+        self.can_ban(user)
         await self._state.http.ban(user.id, self.id, delete_message_days, reason=reason)
 
     async def unban(self, user, *, reason=None):
@@ -1697,6 +1730,8 @@ class Guild(Hashable):
         HTTPException
             Unbanning failed.
         """
+        if not self.me.guild_permissions.ban_members:
+            raise InsufficientPermissions('ban_members')
         await self._state.http.unban(user.id, self.id, reason=reason)
 
     async def vanity_invite(self):
@@ -1724,6 +1759,8 @@ class Guild(Hashable):
         """
 
         # we start with { code: abc }
+        if not self.me.guild_permissions.manage_guild:
+            raise InsufficientPermissions('manage_guild')
         payload = await self._state.http.get_vanity_code(self.id)
 
         # get the vanity URL channel since default channels aren't
